# src/app/docs/swagger-suno-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Suno API Plus Next",
    "description": "Unofficial self-hosted Suno API gateway with account affinity, Studio workflows, Advanced Split, Fast Upload and OpenAI-compatible routes.",
    "version": "2.0.0"
  },
...
  "paths": {
...
    "/api/upload_audio": {
      "post": {
        "summary": "Upload audio for a later Cover request",
        "description": "Measures duration with server-side FFprobe. Files <=30.000s use direct upload; files >30.000s use the synchronous Studio fast uploader (1800s default) with no direct fallback. Send a stable Idempotency-Key for fast uploads.",
        "tags": [
          "default"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 512
            },
            "description": "Stable content-and-metadata operation key. Completed fast uploads replay their result; running keys return 409; submission_unknown keys never auto-retry. Use /api/upload_audio/reconcile for a read-only status check; an explicitly authorized replay must use a new key."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "audio_file"
                ],
                "properties": {
                  "audio_file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "title": {
                    "type": "string",
                    "example": "Uploaded source"
                  },
                  "prompt": {
                    "type": "string",
                    "description": "Optional source description or lyrics metadata."
                  },
                  "image_url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "extension": {
                    "type": "string",
                    "example": "mp3"
                  },
                  "upload_type": {
                    "type": "string",
                    "default": "file_upload"
                  },
                  "pool": {
                    "type": "string",
                    "enum": [
                      "basic",
                      "super",
                      "heavy"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Direct upload or completed Studio project clip.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "upload_method",
                    "upload_id",
                    "clip_id",
                    "duration_seconds",
                    "portable_clip",
                    "account_affinity"
                  ],
                  "properties": {
                    "upload_method": {
                      "type": "string",
                      "enum": [
                        "direct",
                        "studio_fast"
                      ]
                    },
                    "upload_id": {
                      "type": "string"
                    },
                    "upload_id_kind": {
                      "type": "string",
                      "enum": [
                        "studio_project"
                      ]
                    },
                    "studio_project_id": {
                      "type": "string"
                    },
                    "clip_id": {
                      "type": "string"
                    },
                    "song_url": {
                      "type": "string",
                      "format": "uri"
                    },
                    "duration_seconds": {
                      "type": "number"
                    },
                    "portable_clip": {
                      "type": "boolean",
                      "description": "true for Studio-rendered clips that were verified for cross-account lookup, Cover and Extend. This is not Suno is_public."
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "account_affinity": {
                      "type": "string",
                      "description": "Opaque signed value; do not modify it. It is not a Cookie or bearer credential."
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Idempotency conflict/in progress, duplicate catalog audio, or bound account unavailable."
          },
          "422": {
            "description": "Duration could not be determined or Suno definitively rejected the upload."
          },
          "429": {
            "description": "Fast uploader or account pool is busy; submission_state is not_submitted."
          },
          "502": {
            "description": "A Studio project exists but the final clip is uncertain; submission_state is submission_unknown and replay is blocked."
          },
          "503": {
            "description": "Runtime/account unavailable before submission."
          },
          "504": {
            "description": "Fast upload timed out; submission_state reports whether a project was observed."
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "AdminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "suno_admin_session"
      }
    }
  },
...
}